server-fastapi/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.1
pymysql==1.1.0
python-dotenv==1.0.1
sqlalchemy==2.0.28 